# Conflicts:
#	react-native/android/src/main/java/com/ricoh360/thetaclientreactnative/ThetaClientSdkModule.kt
#	react-native/ios/ThetaClientReactNative.mm

+++ react-native/android/src/main/java/com/ricoh360/thetaclientreactnative/ThetaClientSdkModule.kt
@@ -9,9 +9,8 @@ import com.ricoh360.thetaclient.capture.VideoCapture
 import com.ricoh360.thetaclient.capture.VideoCapturing
 import io.ktor.utils.io.core.*
 import kotlinx.coroutines.*
-import java.io.ByteArrayOutputStream
-import kotlin.coroutines.*
 import java.util.*
+import kotlin.coroutines.*
 import kotlin.text.String
 
 class ThetaClientReactNativeModule(
@@ -32,8 +31,10 @@ class ThetaClientReactNativeModule(
   var videoCaptureBuilder: VideoCapture.Builder? = null
   var videoCapture: VideoCapture? = null
   var videoCapturing: VideoCapturing? = null
-  lateinit var theta: ThetaRepository
+  var theta: ThetaRepository? = null
   var listenerCount: Int = 0
+
+  val messageNotInit: String = "Not initialized."
 
   /**
    * add event listener for [eventName]
@@ -76,6 +77,14 @@ class ThetaClientReactNativeModule(
   fun initialize(endpoint: String, config: ReadableMap?, timeout: ReadableMap?, promise: Promise) {
     launch {
       try {
+        theta = null
+        previewing = false
+        photoCaptureBuilder = null
+        photoCapture = null
+        videoCaptureBuilder = null
+        videoCapture = null
+        videoCapturing = null
+
         theta = ThetaRepository.newInstance(
           endpoint,
           config?.let { configToTheta(it) },
@@ -89,11 +98,25 @@ class ThetaClientReactNativeModule(
   }
 
   /**
+   * Returns whether it is initialized or not.
+   * @param promise promise to set result
+   */
+  @ReactMethod
+  fun isInitialized(promise: Promise) {
+    promise.resolve(theta != null)
+  }
+
+  /**
    * getThetaInfo  -  retrieve ThetaInfo from THETA via repository
    * @param promise promise to set result
    */
   @ReactMethod
   fun getThetaInfo(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val info = theta.getThetaInfo()
@@ -131,6 +154,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun getThetaState(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val state = theta.getThetaState()
@@ -188,6 +216,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun listFiles(fileType: String, startPosition: Int, entryCount: Int, storage: String?, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val (fileList, totalEntries) = theta.listFiles(
@@ -226,6 +259,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun deleteFiles(fileUrls: ReadableArray, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     val fileList = mutableListOf<String>()
     for (index in 0..(fileUrls.size() - 1)) {
       fileList.add(fileUrls.getString(index))
@@ -246,6 +284,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun deleteAllFiles(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.deleteAllFiles()
@@ -262,6 +305,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun deleteAllImageFiles(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.deleteAllImageFiles()
@@ -278,6 +326,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun deleteAllVideoFiles(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.deleteAllVideoFiles()
@@ -370,6 +423,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun getOptions(optionNames: ReadableArray, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val optionNameList = mutableListOf<ThetaRepository.OptionNameEnum>()
@@ -399,6 +457,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun setOptions(options: ReadableMap, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val thetaOptions = ThetaRepository.Options()
@@ -422,6 +485,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun getLivePreview(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     fun ByteArray.toBase64(): String = String(Base64.getEncoder().encode(this))
     suspend fun callFrameHandler(packet: Pair<ByteArray, Int>): Boolean {
       if (listenerCount == 0) {
@@ -454,6 +522,9 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun stopLivePreview() {
+    if (theta == null) {
+      throw Exception(messageNotInit)
+    }
     previewing = false
   }
 
@@ -462,6 +533,7 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun getPhotoCaptureBuilder() {
+    val theta = theta ?: throw Exception(messageNotInit)
     photoCaptureBuilder = theta.getPhotoCaptureBuilder()
   }
 
@@ -473,6 +545,9 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun buildPhotoCapture(options: ReadableMap, promise: Promise) {
+    if (theta == null) {
+      throw Exception(messageNotInit)
+    }
     if (photoCaptureBuilder == null) {
       promise.reject(Exception("no photoCaptureBuilder"))
       return
@@ -501,6 +576,10 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun takePicture(promise: Promise) {
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     if (photoCapture == null) {
       promise.reject(Exception("no photoCapture"))
       return
@@ -623,6 +702,7 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun getVideoCaptureBuilder() {
+    val theta = theta ?: throw Exception(messageNotInit)
     videoCaptureBuilder = theta.getVideoCaptureBuilder()
   }
 
@@ -633,6 +713,10 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun buildVideoCapture(options: ReadableMap, promise: Promise) {
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     if (videoCaptureBuilder == null) {
       promise.reject(Exception("no videoCaptureBuilder"))
       return
@@ -661,6 +745,10 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun startCapture(promise: Promise) {
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     if (videoCapture == null) {
       promise.reject(Exception("no videoCapture"))
       return
@@ -686,6 +774,9 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun stopCapture() {
+    if (theta == null) {
+      throw Exception(messageNotInit)
+    }
     videoCapturing?.stopCapture()
   }
 
@@ -695,6 +786,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun getMetadata(fileUrl: String, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val metaData = theta.getMetadata(fileUrl)
@@ -729,6 +825,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun reset(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.reset()
@@ -745,6 +846,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun restoreSettings(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.restoreSettings()
@@ -761,6 +867,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun stopSelfTimer(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.stopSelfTimer()
@@ -785,6 +896,11 @@ class ThetaClientReactNativeModule(
     applyTopBottomCorrection: Boolean,
     promise: Promise,
   ) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val convertedUrl = theta.convertVideoFormats(fileUrl, is4k, applyTopBottomCorrection)
@@ -802,6 +918,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun cancelVideoConvert(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.cancelVideoConvert()
@@ -818,6 +939,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun finishWlan(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.finishWlan()
@@ -834,6 +960,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun listAccessPoints(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val result = Arguments.createArray()
@@ -885,6 +1016,11 @@ class ThetaClientReactNativeModule(
     proxy: ReadableMap?,
     promise: Promise
   ) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.setAccessPointDynamically(
@@ -928,6 +1064,11 @@ class ThetaClientReactNativeModule(
     proxy: ReadableMap?,
     promise: Promise
   ) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.setAccessPointStatically(
@@ -955,6 +1096,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun deleteAccessPoint(ssid: String, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.deleteAccessPoint(ssid)
@@ -986,6 +1132,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun setBluetoothDevice(uuid: String, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val deviceName = theta.setBluetoothDevice(uuid)
@@ -1004,6 +1155,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun getMySetting(captureMode: String, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val options = theta.getMySetting(ThetaRepository.CaptureModeEnum.valueOf(captureMode))
@@ -1025,8 +1181,13 @@ class ThetaClientReactNativeModule(
    * @param optionNames The target shooting mode
    * @param promise promise to set result
    */
-  @ReactMethod
-  fun getMySettingFromOldModel(optionNames: ReadableArray, promise: Promise) {
+   @ReactMethod
+   fun getMySettingFromOldModel(optionNames: ReadableArray, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val optionNameList = mutableListOf<ThetaRepository.OptionNameEnum>()
@@ -1056,6 +1217,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun setMySetting(captureMode: String, options: ReadableMap, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val thetaOptions = ThetaRepository.Options()
@@ -1080,6 +1246,11 @@ class ThetaClientReactNativeModule(
    */
   @ReactMethod
   fun deleteMySetting(captureMode: String, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         theta.deleteMySetting(ThetaRepository.CaptureModeEnum.valueOf(captureMode))
@@ -1094,8 +1265,13 @@ class ThetaClientReactNativeModule(
    * listPlugins - acquires a list of installed plugins
    * @param promise promise to set result
    */
-  @ReactMethod
-  fun listPlugins(promise: Promise) {
+   @ReactMethod
+   fun listPlugins(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
     launch {
       try {
         val result = Arguments.createArray()
@@ -1127,6 +1303,11 @@ class ThetaClientReactNativeModule(
     */
     @ReactMethod
     fun setPlugin(packageName: String, promise: Promise) {
+     val theta = theta
+     if (theta == null) {
+       promise.reject(Exception(messageNotInit))
+       return
+     }
      launch {
       try {
         theta.setPlugin(packageName)
@@ -1144,6 +1325,11 @@ class ThetaClientReactNativeModule(
     */
     @ReactMethod
     fun startPlugin(packageName: String, promise: Promise) {
+     val theta = theta
+     if (theta == null) {
+       promise.reject(Exception(messageNotInit))
+       return
+     }
      launch {
       try {
         theta.startPlugin(packageName)
@@ -1160,6 +1346,11 @@ class ThetaClientReactNativeModule(
     */
     @ReactMethod
     fun stopPlugin(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
      launch {
       try {
         theta.stopPlugin()
@@ -1177,6 +1368,11 @@ class ThetaClientReactNativeModule(
     */
     @ReactMethod
     fun getPluginLicense(packageName: String, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
      launch {
       try {
         val result = theta.getPluginLicense(packageName)
@@ -1193,6 +1389,11 @@ class ThetaClientReactNativeModule(
     */
     @ReactMethod
     fun getPluginOrders(promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
      launch {
       try {
         val result = theta.getPluginOrders()
@@ -1214,6 +1415,11 @@ class ThetaClientReactNativeModule(
     */
     @ReactMethod
     fun setPluginOrders(plugins: ReadableArray, promise: Promise) {
+    val theta = theta
+    if (theta == null) {
+      promise.reject(Exception(messageNotInit))
+      return
+    }
      launch {
       try {
         val pluginList = mutableListOf<String>()
